Refactoring: move low level tasks to utils.py
* approximating payload execution time
* execution of payload
* setting up of data inversion
